+++ ant-api/src/main/java/com/bx/ant/pageModel/DriverFreightRuleQuery.java
@@ -1,11 +1,13 @@
 package com.bx.ant.pageModel;
 
+import java.math.BigDecimal;
+
 /**
  * Created by w9777 on 2017/11/13.
  */
 public class DriverFreightRuleQuery extends DriverFreightRule {
     private Integer weight;
-    private Integer distance;
+    private BigDecimal distance;
 
     public Integer getWeight() {
         return weight;
@@ -15,11 +17,11 @@ public class DriverFreightRuleQuery extends DriverFreightRule {
         this.weight = weight;
     }
 
-    public Integer getDistance() {
+    public BigDecimal getDistance() {
         return distance;
     }
 
-    public void setDistance(Integer distance) {
+    public void setDistance(BigDecimal distance) {
         this.distance = distance;
     }
 }
